+++ CHANGELOG.md
@@ -14,7 +14,6 @@
 
 ### Other Changes
 
-- Added parameter to get API response without svg string in the result. ([#137])
 - Added documentation on “How to add a layer definitions” ([#141])
 - Use concurrency when creating all indicators ([#153])
 - Add and improve tests for API ([#168])
@@ -22,6 +21,7 @@
     - Improve API response schemata by using less logic to create the schemata.
 - Add list indicator/layer combinations for API and CLI([#99])
 - Load indicator from DB will also load its data attributes ([#179])
+- Added parameter to get API response without svg string in the result. ([#137])
 
 ### How to upgrade?